% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/1_calculate_SRratio.R
\name{calculate_SRratio}
\alias{calculate_SRratio}
\title{Create a Data Frame of SRratio}
\usage{
calculate_SRratio(.data, control, treatment, sample_pair, is.logarithmic = NA)
}
\arguments{
\item{.data}{A data frame with columns representing feature and samples.}

\item{control}{A character vector containing column names of control samples.}

\item{treatment}{A character vector containing column names of treatment samples.}

\item{sample_pair}{A data frame with control samples and treatment samples.}

\item{is.logarithmic}{A logical specifying if .data is log-transformed.}
}
\value{
SRratio
}
\description{
SRratio is the gene expression ratios between the control samples and
treatment samples across all combinations within a research project.
}
\examples{
ebg <- expand_by_group(MetadataABA, control_sample, treated_sample, Series)

SRratio <- calculate_SRratio(TranscriptomeABA, "control_sample",
                             "treated_sample", ebg, is.logarithmic = 1)

}
