% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/2_calculate_SRscore.R
\name{calculate_SRscore}
\alias{calculate_SRscore}
\title{Create a Data Frame of SRscore}
\usage{
calculate_SRscore(SRratio)
}
\arguments{
\item{SRratio}{A data frame of SRratio.}
}
\value{
A data frame containing results.
}
\description{
SRscore is score value of genes based expression profiles across different research projects.
SRratio is required to calculate SRscore.
}
\examples{
ebg <- expand_by_group(MetadataABA, control_sample, treated_sample, Series)

SRratio <- calculate_SRratio(TranscriptomeABA, "control_sample",
                             "treated_sample", ebg, is.logarithmic = 1)

head(calculate_SRscore(SRratio))

}
