% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_diffexp.R
\name{get_diffexp}
\alias{get_diffexp}
\title{Obtain the conditions under which differential expression of the specified gene was observed.}
\usage{
get_diffexp(genes, srratio, metadata)
}
\arguments{
\item{genes}{description}

\item{srratio}{description}

\item{metadata}{description}
}
\description{
Obtain the conditions under which differential expression of the specified gene was observed.
}
\examples{
ebg <- expand_by_group(MetadataABA, control_sample, treated_sample, Series)

SRratio <- calculate_SRratio(TranscriptomeABA, "control_sample",
                             "treated_sample", ebg, is.logarithmic = 1)
set.seed(1)
get_diffexp(sample(SRratio$ensembl_gene_id, 1), SRratio, MetadataABA)


}
