% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MetadataABA.R
\docType{data}
\name{MetadataABA}
\alias{MetadataABA}
\title{Metadata of abscisic acid stress microarray dataset in *Arabidopsis thaliana*}
\format{
A data frame with 19 rows and 4 variables:
\describe{
  \item{Series}{Research project ID}
  \item{control_sample}{control sample ID}
  \item{treated_sample}{treatment sample ID}
  \item{treatment}{treatment condition}
}
}
\usage{
MetadataABA
}
\description{
Metadata of abscisic acid stress microarray dataset in *Arabidopsis thaliana*
}
\keyword{datasets}
