% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MetadataHypoxia.R
\docType{data}
\name{MetadataHypoxia}
\alias{MetadataHypoxia}
\title{Metadata of Hypoxia stress RNA-Seq dataset in \emph{Arabidopsis thaliana}}
\format{
A data frame with 29 rows and 4 variables:
\describe{
\item{study_accession}{Research project ID}
\item{run_accession..hypoxia.}{RNA-Seq run accession ID}
\item{run_accession..normoxia.}{RNA-Seq run accession ID}
\item{Treatment}{treatment condition}
}
}
\usage{
MetadataHypoxia
}
\description{
This is metadata of RNA-Seq data that is used in the study by Tamura and Bono.
}
\keyword{datasets}
