% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sample_pair_test.R
\docType{data}
\name{sample_pair_test}
\alias{sample_pair_test}
\title{Test data to create data frames from all combinations between two specified variables within each group using sample data}
\format{
A data frame with 71 rows and 2 variables:
\describe{
\item{control_sample}{Control Sample ID}
\item{treated_sample}{Treated Sample ID}
}
}
\usage{
sample_pair_test
}
\description{
Test data to create data frames from all combinations between two specified variables within each group using sample data
}
\keyword{datasets}
