% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SRGA.R
\docType{data}
\name{SRGA}
\alias{SRGA}
\title{Test data for template matching using sample data}
\format{
A data frame with 1000 rows and 13 variables:
\describe{
\item{ensembl_gene_id}{Ensembl gene ID}
\item{ABA}{SRscore derived from ABA dataset}
\item{Cold}{SRscore derived from cold dataset}
\item{DC3000}{SRscore derived from DC3000 dataset}
\item{Drought}{SRscore derived from drought dataset}
\item{Heat}{SRscore derived from heat dataset}
\item{High-light}{SRscore derived from highlight dataset}
\item{Hypoxia}{SRscore derived from hypoxia dataset}
\item{Osmotic}{SRscore derived from osmotic dataset}
\item{Oxidation}{SRscore derived from oxidation dataset}
\item{Salt}{SRscore derived from salt dataset}
\item{Wound}{SRscore derived from wound dataset}
\item{SYMBOL}{Gene symbol}
}
}
\usage{
SRGA
}
\description{
SRGA is a data frame integrating SRscores across 11 distinct stress conditions(Fukuda et al., 2025).
Since the SR score scale varies by stress type, it was standardized using z-scores.
To reduce file size, only 1,000 genes are included.
By utilizing template matching (Pavlidis and Noble 2001), genes exhibiting similar SRscore patterns under each stress condition can be extracted.
}
\keyword{datasets}
