% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SRratio_test.R
\docType{data}
\name{SRratio_test}
\alias{SRratio_test}
\title{Test data for calculating SRratio using sample data}
\format{
An object of class \code{data.frame} with 1000 rows and 20 columns.
}
\usage{
SRratio_test
}
\description{
A dataframe containing SRratio calculated from \code{TranscriptomeABA}
}
\details{
Column components :

Ensembl gene id + 19 treatment sample id
}
\keyword{datasets}
