% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TPMHypoxia.R
\docType{data}
\name{TPMHypoxia}
\alias{TPMHypoxia}
\title{Hypoxia stress RNA-Seq data in \emph{Arabidopsis thaliana}}
\format{
An object of class \code{data.frame} with 1000 rows and 59 columns.
}
\usage{
TPMHypoxia
}
\description{
This is RNA-Seq data that is used in the study by Tamura and Bono (2022) \url{doi:10.3390/life12071079}.
The quantitative RNA-Seq data, which were calculated as transcripts per million (TPM),
are available at figshare \url{doi:10.6084/m9.figshare.20055086}.
}
\details{
Column components :

Ensembl gene id + 58 sample id (control : 29, treatment : 29)
}
\keyword{datasets}
