% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TranscriptomeABA.R
\docType{data}
\name{TranscriptomeABA}
\alias{TranscriptomeABA}
\title{Abscisic acid stress microarray dataset in \emph{Arabidopsis thaliana}}
\format{
An object of class \code{data.frame} with 1000 rows and 39 columns.
}
\usage{
TranscriptomeABA
}
\description{
This is a gene expression matrix for Arabidopsis under ABA stress conditions.
The first column is the gene ID column, all others are sample ID columns.
The expression data are read as raw data (CEL files) and summarized and normalized by Robust Multi-array Average (RMA).
To keep the file size small, the data is limited to 1,000 genes.
}
\details{
Column components :

Ensembl gene id + 38 sample id (control : 19, treatment : 19)
}
\keyword{datasets}
