% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/0_expand_by_group.R
\name{expand_by_group}
\alias{expand_by_group}
\title{Create a data frame from all combinations between conditions by each research project}
\usage{
expand_by_group(.data, cnt, trt, .by)
}
\arguments{
\item{.data}{A data frame containing at least three variables.}

\item{cnt}{An unquoted expression specifying the column of control conditions.}

\item{trt}{An unquoted expression specifying the column of treatment conditions.}

\item{.by}{An unquoted expression specifying the column of research projects.}
}
\value{
A data frame containing one row for each combination of the specified conditions.
}
\description{
You need to supply a data frame and at least three variables in the dataframe.
This is useful when creating a data frame from all combinations of
control conditions and treatment conditions across research projects.
}
\examples{
ebg <- expand_by_group(MetadataABA,
                       control_sample,
                       treated_sample,
                       Series)

unique_series <- unique(MetadataABA$Series)

lapply(unique_series,
       function(x) subset(ebg, Series == x))


}
