% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/logHNratioHypoxia.R
\docType{data}
\name{logHNratioHypoxia}
\alias{logHNratioHypoxia}
\title{Test data for calculating HNratio using sample data}
\format{
An object of class \code{data.frame} with 1000 rows and 30 columns.
}
\source{
Tamura, Keita, and Hidemasa Bono. 2022. “Meta-Analysis of RNA Sequencing Data of Arabidopsis and Rice Under Hypoxia.” Life 12 (7).
}
\usage{
logHNratioHypoxia
}
\description{
HNscore was calculated to evaluate the changes in the gene expression under hypoxia across the experiments.
A dataframe containing HNratio calculated from \code{TPMHypoxia}
}
\details{
Column components :

Ensembl gene id + 29 treatment sample id
}
\keyword{datasets}
