% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MetadataABA.R
\docType{data}
\name{MetadataABA}
\alias{MetadataABA}
\title{Metadata of abscisic acid stress microarray dataset in \emph{Arabidopsis thaliana}}
\format{
A data frame with 19 rows and 4 variables:
\describe{
\item{Series}{Research project ID}
\item{control_sample}{control sample ID}
\item{treated_sample}{treatment sample ID}
\item{treatment}{treatment condition}
\item{tissue}{tissue name}
}
}
\usage{
MetadataABA
}
\description{
MetadataABA is the metadata for the experimental dataset related to Arabidopsis thaliana under ABA stress conditions.
Metadata are used to define pairs for comparison between the target sample group and the experimental sample group.
}
\keyword{datasets}
