% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TPMHypoxia.R
\docType{data}
\name{TPMHypoxia}
\alias{TPMHypoxia}
\title{Hypoxia stress RNA-Seq data in *Arabidopsis thaliana*}
\format{
A data frame with 1000 rows and 59 variables:
\describe{
  \item{transcript_id}{Transcript ID}
  \item{SRR15606823}{Sample ID}
  \item{SRR15606824}{Sample ID}
  \item{SRR15606825}{Sample ID}
  \item{SRR15606826}{Sample ID}
  \item{SRR15606827}{Sample ID}
  \item{SRR15606828}{Sample ID}
  \item{SRR3139504}{Sample ID}
  \item{SRR3139505}{Sample ID}
  \item{SRR3139506}{Sample ID}
  \item{SRR3139507}{Sample ID}
  \item{SRR3139512}{Sample ID}
  \item{SRR3139513}{Sample ID}
  \item{SRR3139514}{Sample ID}
  \item{SRR3139515}{Sample ID}
  \item{SRR5164962}{Sample ID}
  \item{SRR5164963}{Sample ID}
  \item{SRR5164964}{Sample ID}
  \item{SRR5164965}{Sample ID}
  \item{SRR8206428}{Sample ID}
  \item{SRR8206429}{Sample ID}
  \item{SRR8206430}{Sample ID}
  \item{SRR8206431}{Sample ID}
  \item{SRR8206432}{Sample ID}
  \item{SRR8206433}{Sample ID}
  \item{SRR8206440}{Sample ID}
  \item{SRR8206441}{Sample ID}
  \item{SRR8206442}{Sample ID}
  \item{SRR8206443}{Sample ID}
  \item{SRR8206444}{Sample ID}
  \item{SRR8206445}{Sample ID}
  \item{SRR8234123}{Sample ID}
  \item{SRR8234124}{Sample ID}
  \item{SRR8234125}{Sample ID}
  \item{SRR8234126}{Sample ID}
  \item{SRR8234127}{Sample ID}
  \item{SRR8234128}{Sample ID}
  \item{SRR8234129}{Sample ID}
  \item{SRR8234130}{Sample ID}
  \item{SRR8234131}{Sample ID}
  \item{SRR8234132}{Sample ID}
  \item{SRR8234133}{Sample ID}
  \item{SRR8234134}{Sample ID}
  \item{SRR9034723}{Sample ID}
  \item{SRR9034724}{Sample ID}
  \item{SRR9034725}{Sample ID}
  \item{SRR9034726}{Sample ID}
  \item{SRR9034727}{Sample ID}
  \item{SRR9034728}{Sample ID}
  \item{SRR9034729}{Sample ID}
  \item{SRR9034730}{Sample ID}
  \item{SRR9034731}{Sample ID}
  \item{SRR9034732}{Sample ID}
  \item{SRR9034733}{Sample ID}
  \item{SRR9034734}{Sample ID}
  \item{SRR966478}{Sample ID}
  \item{SRR966479}{Sample ID}
  \item{SRR966482}{Sample ID}
  \item{SRR966483}{Sample ID}
}
}
\usage{
TPMHypoxia
}
\description{
This is RNA-Seq data that is used in the study by Tamura and Bono.
The quantitative RNA-Seq data, which were calculated as transcripts per million (TPM),
are available at figshare (https://figshare.com/articles/dataset/Meta-analysis_of_RNA_Sequencing_Data_of_Arabidopsis_and_Rice_under_Hypoxia/20055086).
}
\keyword{datasets}
