% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TranscriptomeABA.R
\docType{data}
\name{TranscriptomeABA}
\alias{TranscriptomeABA}
\title{Abscisic acid stress microarray dataset in *Arabidopsis thaliana*}
\format{
A data frame with 1000 rows and 39 variables:
\describe{
  \item{ensembl_gene_id}{Ensembl gene ID}
  \item{GSM153922}{Sample ID}
  \item{GSM153923}{Sample ID}
  \item{GSM153924}{Sample ID}
  \item{GSM153925}{Sample ID}
  \item{GSM170896}{Sample ID}
  \item{GSM170897}{Sample ID}
  \item{GSM170923}{Sample ID}
  \item{GSM170931}{Sample ID}
  \item{GSM4012617}{Sample ID}
  \item{GSM4012618}{Sample ID}
  \item{GSM4012619}{Sample ID}
  \item{GSM4012620}{Sample ID}
  \item{GSM4012621}{Sample ID}
  \item{GSM4012622}{Sample ID}
  \item{GSM486916}{Sample ID}
  \item{GSM486917}{Sample ID}
  \item{GSM486918}{Sample ID}
  \item{GSM486928}{Sample ID}
  \item{GSM486929}{Sample ID}
  \item{GSM486930}{Sample ID}
  \item{GSM713249}{Sample ID}
  \item{GSM713250}{Sample ID}
  \item{GSM713251}{Sample ID}
  \item{GSM713255}{Sample ID}
  \item{GSM713256}{Sample ID}
  \item{GSM713257}{Sample ID}
  \item{GSM967132}{Sample ID}
  \item{GSM967133}{Sample ID}
  \item{GSM967140}{Sample ID}
  \item{GSM967141}{Sample ID}
  \item{GSM967148}{Sample ID}
  \item{GSM967149}{Sample ID}
  \item{GSM967156}{Sample ID}
  \item{GSM967157}{Sample ID}
  \item{GSM967164}{Sample ID}
  \item{GSM967165}{Sample ID}
  \item{GSM967172}{Sample ID}
  \item{GSM967173}{Sample ID}
}
}
\usage{
TranscriptomeABA
}
\description{
Abscisic acid stress microarray dataset in *Arabidopsis thaliana*
}
\keyword{datasets}
